Add Javascript code to obtain a GitHub App access token
In order to act as a GitHub App, we need to obtain a so-called
"installation access token", which is like a Personal Access Token for
GitHub Apps, but scoped to a specific repository. To that end, we need
to figure out the "installation ID" of the GitHub App on that specific
repository.

Here are two support functions to make this happen, both making use of
the previously-introduced function to perform GitHub API requests on
behalf of a GitHub App.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>