Add code to perform GitHub REST API calls using an access token
Previously, we introduced functionality to perform GitHub REST API
requests on behalf of a GitHub App. However, this method does not work
for API requests acting on repositories because those requests require
access using an installation access token.

This commit introduces a base function to perform the latter class of
API requests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>