Do make sure that the build did not leave uncommitted changes behind
The strategy used by Git for Windows for years was that the Azure
Pipeline that built the packages and deployed them would then commit the
changes and push directly to the `main` branch.

However, we are changing that model out: We'll use branch deploys going
forward, where packages are built and deployed from PR branches, and
once that is successful, the PR is merged.

That means that all necessary changes, including the `pkgver`
modification, needs to be made up front, not by the same build that
deploys the packages.

Make sure of that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>