Set the check-run to "in progress" when initiating it
The "queued" status is the default, but the "in progress" status
is more appropriate at this stage.

In preparation for maybe eventually driving all of this from a _real_
GitHub App that listens to webhook events and potentially already
creates the check run before triggering the workflow run, let's also
make sure that any non-in-progress check-run is marked as "in progress".

Note: Once a check run has a `conclusion` attached to it, it is no
longer possible to move it back into in-progress. Therefore, in such
instances we have to create new check runs instead of reusing existing
ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>