Support building MSYS packages
There are two locations where Git for Windows maintains definitions for
MSYS packages: in git-for-windows/MSYS2-packages, and the
`git-for-windows-keyring` package in git-for-windows/build-extra.

The main difference compared to MINGW packages is that we need to use
`makepkg` instead of `makepkg-mingw`.

Another difference is that we have to build MSYS packages separately for
both supported architectures (i686 and x86_64, AKA "32-bit" and
"64-bit") because they need to be built in the Git for Windows SDK of
the same architecture.

Finally, MSYS packages do not have a `mingw-w64-` prefix in their
package name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>