Add a workflow to upgrade Git for Windows' packages
This workflow will perform the rather repetitive tasks of upgrading Git
for Windows' packages to newer versions.

With this workflow, we pursue a new strategy to upgrade those packages:
We previously had an Azure Pipeline that edited the `PKGBUILD`, built
the packages, then deployed them to the Pacman repository, and only when
everything worked, the commit(s) with the update would be pushed
directly to `main`.

The new strategy is to instead have a workflow run (this one!) to edit
the `PKGBUILD`, and then open a PR.

A subsequent PR run will verify that the package builds.

Another workflow (the sibling of this one, `build-and-deploy`) will then
"branch-deploy", i.e. build and upload the packages from that PR.

Once everything worked, the PR will be merged.

Hopefully that process will make it much more transparent how the Git
for Windows project is run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>