open-pr: allow acting on behalf of a user
The main use case for this workflow is to be triggered via `/open pr`
commands (see for example
https://github.com/git-for-windows/git/issues/4141#issuecomment-1334538998).

In such a scenario, this workflow is triggered using the installation
access token of the GitForWindowsHelper GitHub App, which does not
really have any identity we can use.

Let's allow specifying explicitly on whose behalf we're running this, so
that the commit author can be configured correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>