build-and-deploy: allow acting on behalf of a user
Just like we changed the `open-pr` workflow to allow for the
GitForWindowsHelper GitHub App to act on a user's behalf, we want the
`build-and-deploy` workflow to be run in the same way.

Let's allow specifying explicitly on whose behalf we're running this, so
that the commit author can be configured correctly.

Also: It is quite common for users not to publish their email addresses
in their GitHub profile, which is totally okay. While already in the
area, let's deal with this in an elegant way.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>