build-and-deploy(i686): rebase the `.dll` base addresses
This is necessary to avoid the dreaded i686 problems described in
https://github.com/git-for-windows/git/wiki/32-bit-issues whose symptoms
can look like this:

      1 [main] perl 5307 child_info_fork::abort: address space needed by 'Cwd.dll' (0x1F0000) is already occupied

For an example where this symptom raises its ugly head see e.g.
https://github.com/git-for-windows/git-for-windows-automation/actions/runs/3747937026/jobs/6364819440

Let's use the common strategy to run the `rebaseall` script. This is a
slightly edited copy of the same step in MSYS2-packages'
`build-and-deploy` workflow:
https://github.com/git-for-windows/MSYS2-packages/blob/6628f78d1359/.github/workflows/build-and-deploy.yml#L60-L76

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>