open-pr: be more lenient when looking for the `component-update` ticket
When opening a Pull Request, we would like to automatically close the
corresponding `component-update` ticket, if there is any.

Unfortunately, it is not _quite_ trivial to search for such tickets. One
frustrating aspect is that some updates have the version number prefixed
with `v` while others do not. And the search on GitHub is a bit strict
about that: if you search for `0.10.1`, you won't find a ticket
mentioning `v0.10.1` and the opposite is also true.

We already started to special-case at least Git LFS, but it's a bit of a
whack-a-mole to do that for all components we track in Git for Windows.

Let's instead search for the version in both forms: both prefixed with
`v` as well as without.

This addresses
https://github.com/git-for-windows/gfw-helper-github-app/issues/18

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>