self-hosted-runners: make Actions Runner path customizable
Some Azure VM types have a temporary disk which is local to the VM and
therefore provides _much_ faster disk IO than the OS Disk (or any other
attached disk).

Hosted GitHub Actions runners also leverage this disk and do their work
in D:/a/_work, so let's use it too if we can. It leads to a ~25% speed
increase when doing heavy IO operations.

Ref: https://learn.microsoft.com/en-us/azure/virtual-machines/managed-disks-overview#temporary-disk
Ref: https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>