self-hosted-runners: do not deallocate by default
We will soon be adding logic that automatically spins up and removes
runners on-demand, so there's no need to deallocate in that case. An
additional benefit of not deallocating is that we can leverage the fast
D:\ drive to install the Runner on. This drive would otherwise be
destroyed as part of the deallocation process.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>