self-hosted-runners: improve approach for Azure ARM parameters
The string we used to provide parameters to Azure ARM was getting
way too long. Let's use a different approach by preparing the string in
the previous step. An additional benefit to this is that we can use
Bash's environment variable expansion, which should be more robust
than the one GitHub provides through ${{ env.XYZ }}.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>