Add a workflow to delete a self-hosted runner VM
Since we're using ephemeral runners, the VMs are no longer used after
they completed a workflow job, and need to be deleted. This GitHub
workflow is designed to perform that action, and it is intended to be
triggered by the GitForWindowsHelper GitHub App in response to receiving
a `workflow_job.completed` event.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>