build-and-deploy: look harder for the 32-bit SDK
The GitHub Action `setup-git-for-windows-sdk` recently learned to detect
when a fast `D:` drive is present, and if so, to use it to check out the
Git for Windows SDK there.

Typically, we do not really care where the SDK is, as its relevant
directories are added to the `PATH` anyway, and therefore any `shell:
bash` step will pick it up without having to hard-code any path.

However, in the i686 flavor of the SDK, we unfortunately have to deal
with the DLL base address problems. We do that by running the
`rebase.exe` helper to ensure that no DLL address ranges clash.

This step cannot use the Bash, because `bash.exe` links to
`msys-2.0.dll`, and therefore _that_ DLL's base address could not be
changed because the DLL is in use.

Therefore we run that step as a PowerShell step, where we _have_ to
hard-code the location, so that it knows where `rebase.exe` is.

Let's just detect whether the SDK is on the `D:` drive or not, and use
the appropriate location in the `rebase` step.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>