build-and-deploy: restrict the `PATH` to MSYS2
On GitHub's hosted runners, there are plenty of things installed. For
example Strawberry Perl. With a `gmake.exe` in the `PATH`. Which is
picked up by `libfido2`'s CMake definition, and the build fails.

Let's prevent that from happening by restricting the `PATH` to the bare
essentials. This commit is a sibling to:
https://github.com/git-for-windows/MSYS2-packages/commit/2733f6da822d

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>