open-pr: avoid using the `setup-git-for-windows-sdk` Action
We actually need something different than the `minimal` SDK (which is
used in git/git's CI builds to build Git and run its test suite): we
need `updpkgsums` and friends.

So far, we abused the Action to initialize the minimal SDK and then
called `curl` plenty a time to download the remaining files.

It would appear that that downloading part ran afoul of rate limits at
least in one instance.

Let's just use a partial clone to begin with.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>