open-pr: cache the git-sdk-64 subset
Caching this instead of re-checking out all the time is a big time
saver. In particular when something failed such as downloading the
source archive, and something in `update-scripts/` needs to be added or
to be adjusted.

Instead of using the `actions/cache` Action as-is, we use the `restore`
and `save` sub-Actions for tighter control (i.e. to avoid caching the
artifact when the clone failed due to problems outside of our control
such as network glitches).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>