build-and-deploy: refresh the installation access token as needed
The build
https://github.com/git-for-windows/git-for-windows-automation/actions/runs/4153007326/jobs/7184357891
as well the build
https://github.com/git-for-windows/git-for-windows-automation/actions/runs/4153007483/jobs/7184358147
failed due to an expiring installation access token: these tokens expire
after an hour, and the step that built the Pacman package took well over
an hour.

To cope with that, remember the expiry date of the token, and after the
Pacman package has been built, refresh it if it is expired (or running
the danger of expiring while the packages are deployed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>