git-artifacts: allow building from an existing tag
So far, we have an implementation of the common flow where the `tag-git`
workflow is started to create "bundle-artifacts" including a Git bundle
with the newly-tagged Git version and another Git bundle with the
updated release notes.

However, for the just-published Git for Windows v2.39.2, we would like
to build the ARM64 flavor of the artifacts, i.e. for an _existing_ tag.

Let's introduce two new inputs (referring to an existing tag and an
existing build-extra revision) that can be used instead of the input
referring to a `tag-git` workflow run, and handle that scenario, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>