open-pr: include `base64` in the SDK subset
We need this tool for pushing (to be precise, to generate the
authorization header).

For the same reason as explained in 8d36b48 (open-pr: avoid caching
error 127, 2023-02-21) where the interplay of two incompatible MSYS2
runtime versions would cause an unhelpful exit code 127, calling Git for
Windows' `base64.exe` from the SDK subset's Bash should be avoided.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>