Distill Check Run update functionality into a separate GitHub Action
One of the main reasons of the existence of the
`git-for-windows-automation` repository is the functionality where the
`build-and-deploy` workflow "mirrors" a Check Run into a repository that
is different from the one that actually runs the workflow.

This allows us to version the workflow definition implementing automated
builds independently of the code that is actually being built.

We will also use this functionality in the `git-artifacts` workflow (and
related workflows).

So let's make the code as reusable as possible by creating a new GitHub
Action specifically for the purpose.

To make things a bit more convenient, the Action maintains a state in a
file in `runner.temp`. This file contains, among other things, the
access token as well as the information when it expires. Since that is
sensitive information and should not be accessible by the build
processes that are run as part of the rest of the workflow, the state
file is encrypted (using the GitHub App's private key that we need to
generate an access token, anyway).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>