tag-git: don't bother setting `HOME`
We're using Linux runners to run the `tag-git` workflow (simply because
they start up quicker than Windows runners), which is a deviation from
the previous process where the tag was created as part of the `Git
Artifacts` Azure Pipeline (which needed to use a Windows runner to
build the artifacts).

This means that we no longer have to set `HOME` and `USERPROFILE`: the
former will be set (unlike on Windows) and the latter does not need to
be set (unlike on Windows).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>