git-artifacts: reword stale description
The `tag-git` workflow originally started out as a literal translation
of the previous `bundle-artifacts` job that was part of the `Git
artifacts` Azure Pipeline.

But since the name `tag-git` made a lot more sense, especially in a
stand-alone GitHub workflow, this is the name with which we ended up,
and this is the name we want to use in the corresponding description.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>