tag-git: mention the tag name in the mirrored Check Run
It is important to surface vital information in the Check Run, such as
the name of the generated tag.

That can help avoid wasting time when an unintended tag name is used: in
that instance we do not need to wait until the Git artifacts are
generated merely to find out that we have to redo the whole thing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>