https-request: offer to upload files, too
When an entire file needs to be uploaded, it is probably not a good idea
to read the contents into memory and pass them as parameter. It is much
better to open a stream and pipe it through.

To support that, we allow to pass a file read stream instead of a string
or an object.

Sadly, we cannot use arbitrary read stream, it has to be a file read
stream, because we need to know the file name and the file length.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>