tag-git: also record the git-for-windows/git commit OID
This information should really be easy to get from the workflow run, but
it will be good to have it recorded in the bundle artifacts.

Technically, it is part of the `git.bundle`, but it is relatively hard
to get to that information. One could of course do

	git clone --bare --filter blob:none https://github.com/git-for-windows/git &&
	git fetch $BUNDLE $TAG_NAME &&
	git rev-parse $TAG_NAME^{commit}

but that requires quite a bit of bandwidth and time.

A really dirty alternative would be to run

	sed '1,/^$/d' < $BUNDLE |
	git unpack-objects -r || true

	TAG_OID="$(sed -n 's/\trefs\/tags\/'"$TAG_NAME"'$//p' < $BUNDLE)"
	git rev-parse $TAG_OID^{commit}

but that is _really_ dirty because the `git unpack-objects` call _will_
fail due to missing objects required to complete the thin pack.

Therefore, the best way to provide the commit OID is simply by writing
it into a new file within `bundle-artifacts.zip`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>