git-artifacts: cache the package in a smarter way
The entire idea of caching the package, once built, is to avoid having
to rebuild the Pacman package just because some _other_ step in the
workflow failed.

But using `actions/cache` as-is won't actually cache the artifact
properly (or worse, it can cache it incompletely when the _build_
failed).

Let's be more careful by using the combination of `actions/cache/save`
and `actions/cache/restore`.

While at it, skip other steps, too, when they are not needed because a
cached package was restored.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>