git-artifacts: allow for `build-extra` to have diverged since `tag-git` ran
The `tag-git` workflow, on which the `git-artifacts` runs for the
various architectures rely, publishes a Git bundle to update the release
notes. To make use of that, the `git-artifacts` workflow needs to check
out the `build-extra` repository and merge those changes.

However, it _does_ happen from time to time that the main branch of
`build-extra` advances between the `tag-git` and the `git-artifacts`
run.

Therefore, we need to tell Git to make a merge commit if necessary.

Which means that we have to have Git committer information ready to go,
and therefore we configure `user.*` in the `artifacts` job when building
an `installer`, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>