Re-implement the "Pacman packages" stage
As part of each Git for Windows release, the `mingw-w64-git` Pacman
packages are built and used to make the installers, MinGits, portable
Gits etc.

This here stage's purpose is to publish those Pacman packages (this can
only be done once the Git for Windows installer is validated manually
and released to the wild).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>