release-git: require the `release` branch to be up to date
We want to imitate Azure Release Pipelines' feature where failed stages
can be edited and then re-deployed _without_ requiring the entire
workflow to be re-started anew when its definition was edited.

To that end, the workflow references composite Actions in the `release`
branch (which is expected to be initially identical to the 'main'
branch). That way, the definitions can be edited and pushed to that
`release` branch, and after that re-running failed jobs will pick up
those fixes.

Let's make sure that we do start with the `release` branch's head commit
being the same as the one the workflow run was started from.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>