git-artifacts: ensure that the `tag-git` workflow run has finished
If building from a tag that was generated in a given `tag-git` workflow
run, we must first make sure that that workflow run finished and
succeeded.

This is particularly important because the plan is to teach the
GitForWindowsHelper GitHub App to no longer trigger the `Git Artifacts`
Azure Pipeline upon `/git-artifacts`, but instead to trigger a `tag-git`
workflow run, wait for it to finish, and then to trigger the
`git-artifacts` workflow runs (one for each architecture supported by
Git for Windows).

And the logic to wait for the workflow run to finish and then to trigger
the `git-artifacts` workflow runs is a bit tricky: the actual `tag-git`
workflow run does not provide the necessary metadata (the
git-for-windows/git commit to which to mirror the Check Run) that would
be accessible by the regular `get-a-workflow-run` GitHub REST API call.

Therefore, the plan is not actually to wait for the `tag-git` _workflow
run_ in `git-for-windows-automation` to complete, but instead for the
`tag-git` _Check Run_ that is mirrored into `git-for-windows/git`. This
Check Run is created and updated explicitly and therefore we could make
sure that the relevant information is included there (the `output.text`
attribute contains a link to the workflow run, including its ID, and the
`head_sha` attribute points to the git-for-windows/git commit to which
the newly generated tag points).

Since that Check Run is marked as completed by the workflow run that is
obviously still running, the webhook event might arrive, and the
`git-artifacts` workflow runs might be triggered, _before_ the `tag-git`
workflow run actually finished. Hence the need to potentially await its
completion.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>