repository-update: use the correct git dir for pushing
It is a bit tricky to work with bare repositories these days: The git
dir has to be specified explicitly.

To support pushing from bare repositories, we therefore need to pass the
`--git-dir` option.

However, in some cases we need non-bare clones because we are doing more
than just merging Git bundles.

In those instances, we have to access the `.git` subdirectory of the
directory into which the repository was cloned.

This logic should have been implemented right away, but instead the
copy/pasted, hard-coded directory name `git` was left in place by
mistake.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>