repository-updates: fix cloning bare repositories
Depending whether we need a bare or a non-bare repository to work with,
we need to pass the `--bare` option to the `git clone` invocation.

The technique that was originally intended to do that used the spread
operator, to insert either no element, or a single element into the
argument list of that `git clone` invocation.

However, due to a programming mistake, the brackets were forgotten and
we instead used the spread operator on a string, adding each character
as an individual argument.

The original intention was to use the spread operator on an empty array
in case a non-empty repository is needed instead of the empty string.

However, in the meantime we also (ab-)use the fact that an empty string
is kind of "false-y" and can thus be used in a conditional, whereas an
empty array would be interpreted as "true-y".

Lucky for us, the empty string is iterable and returns no element, as
intended.

Therefore, we keep using an empty string in case a non-bare repository
is required, and an array with the single element `--bare` otherwise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>