repository-updates: use `--git-dir` or override `cwd`, not both
Earlier, we adjusted the `git fetch` invocation to work in the correct
repository by overriding the working directory.

However, the clone might be bare, in which case we should not assume
that `safe.bareRepository` is set to `all`, i.e. we should always use
`--git-dir` to work with bare repositories. Therefore, that invocation
already had a `--git-dir` (even if it had to be fixed because it was not
adjusted properly from the copy/pasted value).

Now that we use the correct git dir, we actually must not override the
working directory, otherwise the git dir cannot be found...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>