build-and-deploy: use the correct input name `rev`
Technically, the input name `ref` is incorrect: we expect a commit hash
there, not a ref name.

That's why I changed the name while transmogrifying the code to mirror a
Check Run into a separate GitHub Action.

But in the `build-and-deploy` workflow definition, we cannot easily
rename that input because it would require an accompanying change to
`gfw-helper-github-app`, something I could not justify working on while
being busy with reimplementing the Azure Release Pipeline to release Git
for Windows versions as `release-git.yml`.

That is the reason why I forgot to adjust the spot where this input is
passed from the `build-and-deploy` workflow run to the
`check-run-action`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>