repository-updates: merge provided bundle not only for `build-extra`
When releasing Git for Windows v2.40.0, the `MINGW-packages` updates
were not pushed. The reason was the faulty logic that cloned that
repository, then _fetched_ from the bundle without merging, and pushed
(which claimed that everything was up to date).

Obviously, we need to merge the fetched ref!

It gets a bit more complicated in the case of `MINGW-packages` because
we are in a bare repository...

So let's see if we can fast-forward, and if not, create a temporary
worktree to perform the merge, hoping that it works without merge
conflicts ;-)

(In case of merge conflicts, the maintainer who triggered the release
needs to download the bundle manually, merge the changes, resolve the
conflicts, push to `main`, and then re-run the failed job.)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>