callProg/callGit: behave more backticks in shell scripts
The output of programs that is captured in backtick-enclosed commands in
shell scripts is not used verbatim; Instead, it is assumed that the
trailing newline character (if any) should be trimmed. Let's do the same
here.

This does not currently make any difference because no caller actually
looks at the output. But that will change in the next commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>