release-git: update the `release` branch if necessary
The idea of the `release` branch in `git-for-windows-automation` is to
contain composite GitHub Actions that implement the stages of the
`release-git` workflow. That way, the stages can be edited in case they
fail due to changes outside Git for Windows' control before re-running
the respective stage's job in the `release-git` workflow run.

For all of that to work, the `release` branch needs to have the same tip
commit as the `main` branch when the `release-git` workflow runs.

We already have a check that verifies that at the beginning of the
workflow run.

However, if that check fails, the workflow run refuses to continue and
asks the user to synchronize the `release` branch manually. That is
totally unnecessary, though, it's a task that the workflow can perform
itself just as easily. So here is a commit teaching it that trick.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>