git-artifacts: specifically add MinGit's package-versions as artifact
In the Azure Pipeline from which we migrated, this was added
specifically as a build artifact so that the
`versions/package-versions-$ver-MinGit.txt` file could be committed in
the `build-extra` repository.

I missed this when porting the automation over to a GitHub workflow.
This is part one of the fix.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>