repository-updates: commit the `package-versions` files
The `build-extra` repository tracks the package versions that went into
official Git for Windows releases, listing both the full set as well as
the subset from which MinGit was built.

I had missed this when porting the release automation from the Azure
Release Pipeline to the GitHub workflow.

This is part two of the fix (part one was done in the `git-artifacts`
workflow where the MinGit flavor of the `package-versions` file is now
offered up as a part of the `mingit` build artifact).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>