check-run-action: use correct owner/repository values
In case the installation token is no longer valid, we need to generate
a new one. To that end, we store the information about the owner and
the repository in the state file.

However, the current code has a bug where it ignored the stored values
and expected the owner/repository inputs to be non-empty at all times.
This caused a failure when building v2.41.0-rc0's artifacts in
https://github.com/git-for-windows/git-for-windows-automation/actions/runs/5001089177/jobs/8959839096#step:16:92
with this error message:

  FAILED GitHub REST API call!
  {
    port: 443,
    hostname: 'api.github.com',
    method: 'GET',
    path: '/repos/git-for-windows/undefined/installation',
    headers: {
      'User-Agent': 'GitForWindowsHelperApp/0.0',
      Accept: 'application/json',
      Authorization: '***'
    }
  }

Let's avoid this problem by using the previously-stored information
as needed (and as intended).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>