tag-git: avoid potentially-uninitialized environment variable
The `CREATE_CHECK_RUN` environment variable should be set to "false"
when debugging things, to avoid cluttering the target repository with
bogus runs.

However, in the common case it does not need to be undefined. It can be
defined to "true", which not only unpuzzles the linter but also the
human reader.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>