repository-updates: fetch missing Git objects as needed
It is quite possible for a single-branch, shallow clone with depth 50 to
miss a couple of Git objects that are required to complete the Git
bundle from which to merge the updates.

By pretending to be in a partial clone, we let Git fetch those missing
Git objects as needed. Easy peasy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>