release-git: handle missing `release` branch gracefully
We already have code to fast-forward a `release` branch that is not up
to date.

By mistake, I _deleted_ the `release` branch, however: the repository
settings are such that merging a PR deletes its "head" branch, and I
had merged the `release` branch into the `main` branch. Deleting the
branch broke the logic to fast-forward `release`.

Let's handle missing `release` branches the same as out-of-date
branches, and try to push the current `main` to the `release` branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>