Cope with funky out-of-band releases' file naming scheme
Regular Git for Windows releases are based on upstream Git releases, and
follow the version number pattern 2.MAJOR.MINOR.

For Git for Windows releases that are _not_ based on upstream Git
releases ("out-of-band releases"), a number is appended in parentheses,
starting with the number 2.

The intention was to write corresponding `announcement` and
`release-notes` files using a more file-system-friendly pattern that
simply appends a dot and that number. This is what the `release-git`
workflow expects.

However, in the first out-of-band release since switching Git for
Windows' release automation from Azure Pipelines to GitHub workflows,
it occurs that the announcement and release note files' names are
written unexpectedly with parentheses.

Let's just apply Postel's Law here and just rename those files to the
expected kind of file names.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>