release: explicitly mark full releases as the latest release
According to the REST API documentation at
https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#create-a-release
the `make_latest` should default to `true`.

However, when releasing Git for Windows v2.42.0 this was not so. I had
to mark it as "latest" manually, and I only realized this because the
`gitforwindows.org` update would fail (because it would still pick up
v2.41.0(3) as the latest release).

Let's just make it explicit that we want to mark it as the latest
version (at least as long as we're not publishing a pre-release).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>