git-artifacts: be more patient while waiting for `tag-git` to finish
It is important for the `git-artifacts` job to wait for the `tag-git`
job to finish because it requires the latter's build artifacts, yet it
is started before the latter actually finishes.

However, polling every second, up to 30 times, is way too impatient.
There are sometimes glitches that make `tag-git` take longer to spin
down, and we can just look every 10 seconds, it does not make much
difference to the overall run time, and for good measure let's wait up
to ~10 minutes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>