git-artifacts: reset `pkgrel=1` between releases
In case some update was necessary between releases, we need to increment
the `pkgrel` so that we can deploy new `mingw-w64-git` package versions.

However, the next release should reset that `pkgrel` back to 1.

Noticed by Matthias Aßhauer in
https://github.com/git-for-windows/git/pull/4692#issuecomment-1817775282.

The script that is used to build `mingw-w64-git` has been adjusted
accordingly in https://github.com/git-for-windows/build-extra/pull/534
and with this change, we use the new `--reset-pkgrel` option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>