build-and-deploy: use a better way to ensure that all packages are updated
When `pacman -Syyu` (whose job is to update the packages in the Git for
SDK) updates the MSYS2 runtime in a Bash step that uses said MSYS2
runtime, the step will stop and fail.

A better way to run that command is via PowerShell, and an even better
way is to use the same script that the `sync` automation of the
`git-sdk-*` repositories already uses.

So let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>