ci: make open-pr's SDK subset initialization reusable
One part of the `open-pr` workflow is concerned with initializing a
subset of Git for Windows' SDK that is able to run `pacman` and friends.

We are about to introduce another workflow that needs this, so let's
extract that code into its own composite GitHub Action.

This commit is best viewed with `--color-moved
--color-moved-ws=allow-indentation-change`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>