self-hosted-runner: allow running in private repositories
GitHub's documentation provides a stern warning against registering
self-hosted runners on public repositories:
https://docs.github.com/en/actions/hosting-your-own-runners/managing-self-hosted-runners/about-self-hosted-runners#self-hosted-runner-security

To counter that, we specifically spin up ephemeral self-hosted runners
in `git-for-windows/git-for-windows-automation` and have automation to
prevent unauthorized people from trying to play games with our runners.

However, for testing in separate repositories, this strategy is utterly
inconvenient. And unnecessary, when running in a private repository
anyway. Except that we need to have a public URL for the post-deployment
script. So let's work around that by hard-coding the CI token into that
URL. This should be good enough, especially when we scrub the token from
the logs (manually, if necessary).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>