self-hosted-runner: document two more required secrets
To obtain the token that is necessary to register the runner, we need
some sort of credentials, and the way we do that is by using the
GitForWindowsHelper GitHub App's credentials. These have to be generated
from the App ID and the private key, therefore we require them as
secrets, which has previously not been documented.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>