build-and-deploy: stop GPG agent before trying to remove GPG files
In the `build-and-deploy` GitHub workflow, we try to clean up the GPG
files before finishing the job, but it seems that as of GPG v2.4.x, we
need to stop the GPG agent before doing that. Example:

  https://github.com/git-for-windows/git-for-windows-automation/actions/runs/7906810424/job/21582488405#step:20:20

So let's do that, and while at it, also make sure that the `dirmngr`
(GPG's daemon that takes care of talking to keyservers) is stopped, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>