github-release: in addition to the tag, also push `main`
The idea is always to update the `main` branch after the Git for Windows
release is done, to close the PR.

So far, this is done manually. But it's better to automate this. Because
then I cannot make mistakes like merging via the `Merge` button (as I
had done for v2.44.0-rc1, which complicated subsequent rebases) instead
of pushing to `main`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>