Add a GitHub workflow to prepare the branches for embargoed Git for Windows versions
When embargoed Git for Windows versions (or backport versions) need to
be prepared, a couple of branches need to be created (internally
nicknamed "time-traveling branches"), to pick back up at the precise
revisions from where the preceding version was built. That way, the
embargoed version will only differ in the fixed Git artifacts, and avoid
shipping with updates in other components that might cause risk of
regressions.

This new workflow prepares these branches.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>