git-artifacts: optionally use release branches
For embargoed Git for Windows versions, we already have a
well-established paradigm where the `git-sdk-32`, `git-sdk-64`,
`build-extra` and `MINGW-packages` worktrees are not initialized with
the current state of the `main` branch, but with the current state of
the `git-<version>` branches, the "release branches".

Let's teach the `git-artifacts` workflow to optionally use the same
paradigm.

This supersedes the strategy we established earlier when building from a
tag, where the `build-extra` revision to use could be specified; This
strategy was incomplete, as it left out e.g. `MINGW-packages` which most
likely _also_ wanted to be at a specific revision rather than blindly
using the tip commit of the `main` branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>