git-artifacts: support building MinGit backports from tags
When building MinGit backports from tags (identified by specifying
"mingit" or "mingit mingit-busybox" as `artifacts` value), we need to
use a different default release branch name, e.g.
`mingit-2.43.x-releases`.

While at it, also adjust the check run name to reflect that this is not
a regular Git for Windows build but a MinGit one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>