Let `mergeBundle()` handle tag refs
In constrast to `build-extra` and `MINGW-packages`, the `git` bundle
contains not a branch but a tag.

In preparation for letting `pushRepositoryUpdate()` handle the `git`
repository, let's handle that fetch to store the tag in the local
repository, not just the `FETCH_HEAD`.

This requires a couple of other changes, too, as a tag cannot be checked
out directly (only as a "detached HEAD").

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>