Add a function to mirror embargoed GitHub Releases to public ones
In preparation for a full workflow that mirrors an embargoed GitHub
Release to a GitHub Release in the public repository, here is a Node.JS
function that mirrors an embargoed GitHub Release from a private
repository to a public one.

Releases marked as pre-releases will be published as pre-releases, too.
The release that is slated to be marked as latest upon publishing should
_not_ be marked as pre-release (this `prerelease == false` is the
indicator used by this function).

This is important because there are a couple of additional steps that
need to be taken for the latest release, such as updating the website
and the `./download-stats.sh` in `build-extra`. Also, the NuGet packages
need to be published only for the latest release (and not, say, for
MinGit backports). These steps are not yet implemented, and will be
added over the next commits.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>