Add a workflow to break a pacman upload lease
To prevent parallel uploads of pacman packages from interfering with
each other Git for Windows uses Azure blobs leases without a timeout.
Sometimes it's necessary to manually break such a lease if a job
acquired a lease and failed to release it. We can do that via the
pacman-helper script[1], but having it available as a GitHub
workflow or eventually a slash command makes this easier to dispatch.

This fixes https://github.com/git-for-windows/git-for-windows-automation/issues/40

[1] https://github.com/git-for-windows/build-extra/pull/488

Signed-off-by: Matthias Aßhauer <mha1993@live.de>