github-release: do create a discussion post for each new Git for Windows version
This was the intention of 0718129 (release: create a discussion for
every Git for Windows release, 2024-07-02), but it changed the call that
creates the draft release.

Apparently the intention to create a discussion post is _not_ stored
when creating a draft release, though, sadly. Therefore we need to state
that intention again when we do publish the release.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>