cleanup-self-hosted-runner: order the conditional blocks from small to large
In general, if..else..fi constructs are more readable when the first
block is shorter than the second, i.e. to handle the "easy" things
first; According to Cognitive Load Theory, this uses less mental
resources when wrapping our head around the code.

In this instance, there is an even better reason: The next commit will
add _another_ condition, and having it in this order will avoid adding
another nesting level:

	if too young
		skip
	else if busy
		skip
	else
		force delete VM
	fi

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>