cleanup-self-hosted-runner: avoid deleting VMs that are busy
So far, we avoid deleting VMs when they are too young.

But we can do better than that: since the VMs are intended to host an
ephemeral runner, and that runner is registered with the current
repository, we can have a look whether GitHub says that the runner is
busy. If it is, well, let's leave it a-running!

Since this query requires a larger scope than the standard GitHub
workflows' `GITHUB_TOKEN` provides, we need to authenticate as the
GitForWindowsHelper GitHub App to do that. This is too complicated,
unfortunately, in a GitHub workflow (at least without the risk of using
non-official GitHub Actions), therefore we use the newly-added shell
script to authenticate the GitHub CLI as a GitHub App (which of course
now requires us to check out the repository).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>