cleanup-self-hosted-runner: clean up idle runners earlier
The original idea was to wait at least 3 hours before deleting a VM, to
avoid interfering with long-running workflows.

Now that we have a way to query whether the runner _is_ busy, we do not
need to wait that long. Basically, we only need to ensure that the
runner had a chance to register itself and to then pick up the job.
Typically, this takes around 6-7 minutes, so waiting for an hour is
probably even a tad conservative.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>