cleanup-self-hosted-runner: use workflow commands for better mark-up
When logging output that we're deleting a VM, it should be marked a bit
more prominently. Let's mark it as a warning:
https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/workflow-commands-for-github-actions#setting-a-warning-message

This marks it in red in the output.

Conversely, we want to mention when we skipped the deletion, and to make
those messages stick out a bit (but less than warnings), let's use:
https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/workflow-commands-for-github-actions#setting-a-notice-message

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>