github-release: remove v2.46.1 work-around
The deployment of the v2.46.1 release originally did not work. The
reason was that the GitForWindowsHelper GitHub App lacked the
permissions to create Discussions.

This has been fixed in the meantime, the deployment was completed, and
now we can remove the work-around that allowed us to resume the failed
deployment.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>