CI: add simple Azure Login workflow
We currently depend on the azure/login GitHub Action in various places. It simply runs an "az login" and "az account set" behind the scenes, and has support for plenty of other scenarios.

As we're moving away from azure/cli in more and more places, we could also replace the login script with our own. Let's create a Composite Workflow to achieve this goal.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>