build-and-deploy: work around curl v8.11.0 bug
In order to deploy the bug fix for curl, we need to work around the bug
that this fix addresses... because the bug breaks the deployment of all
MSYS2-packages, including `libcurl`.

Let's work around this by downgrading `libcurl` specifically for this
deployment.

See https://github.com/git-for-windows/MSYS2-packages/pull/199 for more
details.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>