post-deployment-script: fix regex for sha256
There was a bug in the regex that finds the sha256 hash corresponding to the Git binary. Doing things like [Regex]::Escape directly in a string interpolation, apparently isn't supported by PowerShell

This has surprisingly worked until now, because `Git-2.47.1-64-bit.exe` was always the first binary in the list of assets. Now that we're getting the native `-arm64.exe` binary, the bug in the logic got exposed.

This commit fixes the bug by assigning the escaped asset name to a dedicated variable first.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>